// src/types.d.ts
// Generated by https://quicktype.io

import { person, user } from "@prisma/client";

export interface NewsResponse {
  status: string;
  totalResults: number;
  articles: Article[];
}

export interface Article {
  source: Source;
  author: null | string;
  title: string | null;
  description: string | null;
  url: string;
  urlToImage: string | null;
  publishedAt: string | null;
  content: null | string;
}

export interface Source {
  id: null | string;
  name: string;
}

interface accountRequest extends user, person {}